+++ src/sphere/sphere.rs
@@ -5,36 +5,36 @@ use crate::Vec3;
 
 pub(crate) struct Sphere {
     pub center: Vec3,
-    pub radius: f32,
+    pub radius: f64,
 }
 
 impl Sphere {
-    pub fn hits_ray(self, r: &Ray) -> f32 {
+    pub fn hits_ray(self, r: &Ray) -> f64 {
         let oc = r.origin - self.center;
         let a = r.direction.length_squared();
         let half_b = Vec3::dot(oc, r.direction);
         let c = oc.length_squared() - self.radius * self.radius;
         let discriminant = (half_b * half_b) - (a * c);
-        if discriminant < 0f32 {
+        if discriminant < 0f64 {
             -1.0
         } else {
             (-half_b - discriminant.sqrt()) / a
         }
     }
-    pub fn new(center: Vec3, radius: f32) -> Sphere {
+    pub fn new(center: Vec3, radius: f64) -> Sphere {
         Sphere { center, radius }
     }
 }
 
 impl Hittable for Sphere {
-    fn hit(&self, ray: &Ray, t_min: f32, t_max: f32, hit: &mut HitRecord) -> bool {
+    fn hit(&self, ray: &Ray, t_min: f64, t_max: f64, hit: &mut HitRecord) -> bool {
         let oc = ray.origin - self.center;
         let a = ray.direction.length_squared();
         let half_b = Vec3::dot(oc, ray.direction);
         let c = oc.length_squared() - self.radius * self.radius;
 
         let discriminant = half_b * half_b - a * c;
-        if discriminant < 0_f32 {
+        if discriminant < 0_f64 {
             return false;
         };
         let sqrtd = discriminant.sqrt();
